Ultimos retoques

+++ src/main/java/com/alkemy/desafioDisney/Servicio/impl/PeliculaServicioImpl.java
@@ -3,6 +3,7 @@ package com.alkemy.desafioDisney.Servicio.impl;
 import com.alkemy.desafioDisney.Dto.*;
 import com.alkemy.desafioDisney.Entidad.PelioSerie;
 import com.alkemy.desafioDisney.Entidad.Personaje;
+import com.alkemy.desafioDisney.Enum.Errors;
 import com.alkemy.desafioDisney.Exception.ParamNotFound;
 import com.alkemy.desafioDisney.Mapper.GeneroMapper;
 import com.alkemy.desafioDisney.Mapper.PeliculaMapper;
@@ -10,12 +11,12 @@ import com.alkemy.desafioDisney.Mapper.PersonajeMapper;
 import com.alkemy.desafioDisney.Repositorio.PeliculaRepositorio;
 import com.alkemy.desafioDisney.Repositorio.Specification.PeliculaSpecification;
 import com.alkemy.desafioDisney.Servicio.PeliculaServicio;
+import com.alkemy.desafioDisney.Utils.Utils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
 import java.text.ParseException;
-import java.text.SimpleDateFormat;
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
 import java.util.Optional;
 
@@ -43,14 +44,15 @@ public class PeliculaServicioImpl implements PeliculaServicio {
 
     public PelioSerieDTO guardarPelicula(PelioSerieDTO dto) throws ParseException {
         PelioSerie entidad = peliculaMapper.peliDTO2Entidad(dto, true);
-        GeneroDTO generoDTO = generoServicio.guardarGenero(dto.getGeneroDTO());
-        entidad.setGenero(generoMapper.generoDTO2Entidad(generoDTO));
         PelioSerie entidadGuardada = peliculaRepositorio.save(entidad);
         return peliculaMapper.peliEntidad2DTO(entidadGuardada, true);
     }
 
-    public void borrar(String id) {
-        peliculaRepositorio.deleteById(id);
+    public void borrar(String id) throws ParamNotFound{
+        PelioSerie peli = this.buscarPorId(id);
+        if (peli != null){
+            peliculaRepositorio.deleteById(id);
+        }
     }
 
     public PelioSerieDTO modificarPelicula(String id, PelioSerieDTO dtoNuevo, boolean cargarPersonaje) throws ParseException, ParamNotFound {
@@ -59,11 +61,8 @@ public class PeliculaServicioImpl implements PeliculaServicio {
         busqueda.setTitulo(dtoNuevo.getTitulo());
         busqueda.setCalificacion(dtoNuevo.getCalificacion());
         busqueda.setImagen(dtoNuevo.getImagen());
-        busqueda.setGeneroId(dtoNuevo.getGeneroDTO().getId());
-
-        SimpleDateFormat fechaFormat = new SimpleDateFormat("dd/MM/yyyy");
-        Date fetch = fechaFormat.parse(dtoNuevo.getFechaCreacion());
-        busqueda.setFechaCreacion(fetch);
+        busqueda.setGeneroId(dtoNuevo.getGeneroId());
+        busqueda.setFechaCreacion(Utils.stringToDate(dtoNuevo.getFechaCreacion()));
 
         if (cargarPersonaje){
             List<Personaje> personajes = new ArrayList<>();
@@ -79,8 +78,10 @@ public class PeliculaServicioImpl implements PeliculaServicio {
 
     public PelioSerie buscarPorId(String id) throws ParamNotFound {
         Optional<PelioSerie> result = peliculaRepositorio.findById(id);
+
         if (!result.isPresent()) {
-            throw new ParamNotFound("Pelicula con id: " + id + " no fue encontrado");
+            Errors errors = Errors.PELICULA_NOT_FOUND;
+            throw new ParamNotFound(errors.getMessege());
         }
         return result.get();
     }
